requirements.txt
numpy==1.26.4
pandas==2.2.2
flask==3.0.3
scikit-learn==1.4.2
seaborn==0.13.2
pymongo==4.6.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
Flask-Session==0.6.0
gunicorn==21.2.0
plotly==5.19.0
